create-kubeconfig: change the argument to specify the serviceaccount (#4)
This commit changes the argument from serviceaccount token to the
serviceaccount name. With this change, we don't need to know the
serviceaccount token in advance. This is breaking change.

```
$ ./create-kubeconfig
Usage: ./create-kubeconfig SERVICEACCOUNT [kubectl options]

This script creates a kubeconfig to access the apiserver with the specified serviceaccount and outputs it to stdout.
```